requirements.txt
Pillow>=10.0.0
google-generativeai>=0.3.0
pandas>=2.0.0
openai>=1.40.0

